Remove error icon from binding advanced options header
